In addition, removed re-raising of OSError in mainframe.py

git-svn-id: https://bpbible.googlecode.com/svn/trunk@84 20523bf7-2b47-0410-960a-a116b6c96ec9

+++ util/search_utils.py
@@ -7,22 +7,36 @@ arch.search failed
 import cPickle
 import os
 import config
+import gzip
+from configmgr import config_manager
 
+
+search_config = config_manager.add_section("Search")
+search_config.add_item("zip_indexes", False, item_type=bool)
 
 def WriteIndex(index, path = config.index_path):
-	f = open("%s%s.idx"	% (path, index.version), "w")
+	if search_config["zip_indexes"]:
+		f = gzip.GzipFile("%s%s.idxz" % (path, index.version), "w",
+				compresslevel=5)
+	else:
+		f = open("%s%s.idx" % (path, index.version), "wb")
+
 	cPickle.dump(index, f)
 
 def ReadIndex(version, path = config.index_path):
-	f = open("%s%s.idx" % (path, version))
+	if os.path.exists("%s%s.idxz" % (path, version)):
+		f = gzip.GzipFile("%s%s.idxz" % (path, version))
+	else:
+		f = open("%s%s.idx" % (path, version), "rb")
+	
 	return cPickle.load(f)
 
 def IndexExists(version, path = config.index_path):
-	return os.path.exists("%s%s.idx" % (path, version))	
+	return os.path.exists("%s%s.idx" % (path, version)) or \
+			os.path.exists("%s%s.idxz" % (path, version))
 
 def DeleteIndex(version, path = config.index_path):
-	#try:
-	os.remove("%s%s.idx" % (path, version))
-	#except Exception, e:
-	#	return e
-
+	if os.path.exists("%s%s.idx" % (path, version)):
+		os.remove("%s%s.idx" % (path, version))
+	if os.path.exists("%s%s.idxz" % (path, version)):
+		os.remove("%s%s.idxz" % (path, version))
